+++ desktop/src-tauri/build.rs
@@ -1,0 +1,6 @@
+// Copyright (C) 2026 Nik m (@mazurovn)
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+fn main() {
+    tauri_build::build()
+}